minificpp-34 Refactoring.
